Moved Raid-Leader Common functions.
